requirements.txt
streamlit>=1.32.0
pymysql>=1.1.0
pandas>=2.0.0
plotly>=5.18.0
Pillow>=9.0.0
